+++ inc/common/actions_media.php
@@ -1,5 +1,19 @@
 <?php
 
+/**
+ * Allow SVG files 
+ * 
+ */ 
+function add_svg_mime_types($mimes) {
+	$mimes['svg'] = 'image/svg+xml';
+	return $mimes;
+}
+
+function wptuning_add_svg_mime_types() {
+	add_filter('upload_mimes', 'add_svg_mime_types');
+}
+
+
 function wptuning_add_medium_large() {
 	add_filter( 'image_size_names_choose', 'add_medium_large' );
 }